backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
channels>=4.0
daphne>=4.0
channels-redis>=4.1
django-cors-headers>=4.3
psycopg2-binary>=2.9
Pillow>=10.0